% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screens.R
\name{base_stats}
\alias{base_stats}
\title{Screen stocks with a "base" patter}
\usage{
base_stats(data, ticker, trailing_days = 8 * 21)
}
\arguments{
\item{data}{List of price data}

\item{ticker}{ticker selected}

\item{trailing_days}{Number of days of the base}
}
\description{
The idea is to look for stocks in a base for the past few months.
   I also compute volatility statistics that are interesting to filter stocks further.
}
