% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calendar_returns}
\alias{calendar_returns}
\title{Compute Calendar returns at the monthly and yearly level.}
\usage{
calendar_returns(rets, digits = 3, percent = FALSE)
}
\arguments{
\item{rets}{Daily Return xts object for the asset}

\item{digits}{Number of digits}

\item{percent}{Use Percents or decimals?}
}
\description{
Compute Calendar returns at the monthly and yearly level.
}
