% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{compute_breadth}
\alias{compute_breadth}
\title{compute breadth indicator}
\usage{
compute_breadth(data, tickers, threshold = 0.03)
}
\arguments{
\item{data}{List of price data}

\item{tickers}{Character vector of tickers}

\item{threshold}{Percent difference between the EMA-10 and EMA-50 to consider a stock is in an uptrend}
}
\description{
compute breadth indicator
}
