% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_fundamentals}
\alias{get_fundamentals}
\title{Gets fundamental data}
\usage{
get_fundamentals(ticker, element)
}
\arguments{
\item{ticker}{Stock Ticker}

\item{element}{Element}
}
\description{
Gets fundamental data
}
