% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_price_data}
\alias{get_price_data}
\title{Downloads price data}
\usage{
get_price_data(tickers, from = "1990-01-01", to = NULL)
}
\arguments{
\item{tickers}{Character vector of tickers}

\item{from}{Starting date}

\item{to}{end date, defaults to the current date if it's NULL.}
}
\description{
Downloads price data
}
