% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_start_dates}
\alias{get_start_dates}
\title{Get first date of each ticker}
\usage{
get_start_dates(data)
}
\arguments{
\item{data}{xts object with stock data}
}
\description{
Get first date of each ticker
}
