% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{KeltnerBands}
\alias{KeltnerBands}
\title{Compute Keltner Bands}
\usage{
KeltnerBands(x, n.ema = 21, n.atr = 21, multiplier = 2, maType = "EMA")
}
\arguments{
\item{x}{xts price data}

\item{n.ema}{Period of the EMA}

\item{n.atr}{Period of the ATR}

\item{multiplier}{Used to define the width of the band}

\item{maType}{Type of moving average}
}
\description{
Compute Keltner Bands
}
