% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{list_to_returns}
\alias{list_to_returns}
\title{Compute returns of a list of price data}
\usage{
list_to_returns(data)
}
\arguments{
\item{data}{List of stocks price data}
}
\description{
Compute returns of a list of price data
}
