% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screens.R
\name{long_trend_high_momentum}
\alias{long_trend_high_momentum}
\title{Long term trend following strategy based on stocks with momentum}
\usage{
long_trend_high_momentum(data, ticker, n_return = 200)
}
\arguments{
\item{data}{List of price data}

\item{ticker}{ticker selected}

\item{n_return}{Number of days to compute a "long term" return}
}
\description{
Long term trend following strategy based on stocks with momentum
}
