% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screens.R
\name{long_trend_low_vol}
\alias{long_trend_low_vol}
\title{Long term trend following strategy based on stocks low volatility}
\usage{
long_trend_low_vol(data, ticker, n_return = 200)
}
\arguments{
\item{data}{List of price data}

\item{ticker}{ticker selected}

\item{n_return}{Number of days to compute a "long term" return}
}
\description{
Long term trend following strategy based on stocks low volatility
}
