% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screens.R
\name{mean_reversion}
\alias{mean_reversion}
\title{Mean reversion screen}
\usage{
mean_reversion(data, ticker)
}
\arguments{
\item{data}{List of price data}

\item{ticker}{ticker selected}
}
\description{
Mean reversion screen
}
