% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Most common performance metrics}
\usage{
performance(
  ret_mat,
  date_range = NULL,
  tickers = NULL,
  combine = NULL,
  index = "QQQ",
  plot = TRUE,
  print_cor = FALSE,
  print_tbl = FALSE
)
}
\arguments{
\item{ret_mat}{The daily returns matrix}

\item{date_range}{A string with an xts style date-range to filter data}

\item{tickers}{Character vector with tickers that will be part of the result}

\item{combine}{Character vector of tickers to combine in an equal weight portfolio. Results are summarized as Portfolio}

\item{index}{Benchmark index to compute CAPM Beta and Alpha}

\item{plot}{Plot the cumulative return of all assets}

\item{print_cor}{Print the correlation matrix of the assets defined in ticker}

\item{print_tbl}{Prints the summary performance table}
}
\description{
Most common performance metrics
}
