% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance_ticker}
\alias{performance_ticker}
\title{Most common performance metrics for each individual ticker}
\usage{
performance_ticker(data, ticker, date_range = NULL, index = "SPY")
}
\arguments{
\item{data}{Daily price data}

\item{ticker}{Selected ticker}

\item{date_range}{A string with an xts style date-range to filter data}

\item{index}{Benchmark index to compute CAPM Beta and Alpha}
}
\description{
Most common performance metrics for each individual ticker
}
