% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_fund_data}
\alias{read_fund_data}
\title{Reads fundamental data and corrects some errors}
\usage{
read_fund_data(fund_data_file, ticker_lkp)
}
\arguments{
\item{fund_data_file}{File}

\item{ticker_lkp}{Tickers look up table}
}
\description{
Reads fundamental data and corrects some errors
}
