% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screens.R
\name{weekly_rotation}
\alias{weekly_rotation}
\title{Long term trend following strategy based on weekly rebalancing}
\usage{
weekly_rotation(data, ticker, n_return = 200)
}
\arguments{
\item{data}{List of price data}

\item{ticker}{ticker selected}

\item{n_return}{Number of days to compute return.}
}
\description{
Long term trend following strategy based on weekly rebalancing
}
