% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{which_col}
\alias{which_col}
\title{Find which col to adjust}
\usage{
which_col(x, str = "Adjusted")
}
\arguments{
\item{x}{xts price data}

\item{str}{String to look for}
}
\description{
Find which col to adjust
}
