% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{xts_to_dt}
\alias{xts_to_dt}
\title{Convert an xts object to data.table}
\usage{
xts_to_dt(df_xts)
}
\arguments{
\item{df_xts}{xts object with stock data}
}
\description{
Convert an xts object to data.table
}
